Move collectContentPre to shared and enable for server
