+++ ep_headings/ep.json
@@ -8,11 +8,12 @@
         "postAceInit": "ep_headings/static/js/index",
         "aceInitialized": "ep_headings/static/js/index",
         "aceAttribsToClasses": "ep_headings/static/js/index",
-        "collectContentPre": "ep_headings/static/js/index",
+        "collectContentPre": "ep_headings/static/js/shared",
         "aceRegisterBlockElements": "ep_headings/static/js/index"
       },
       "hooks": {
-        "eejsBlock_editbarMenuLeft": "ep_headings/index"
+        "eejsBlock_editbarMenuLeft": "ep_headings/index",
+        "collectContentPre": "ep_headings/static/js/shared"
       }
     }
   ]
